+++ dashboard/app/src/components/header/mobile-menu/MobileMenu.css
@@ -1,0 +1,5 @@
+.mobile-menu button {
+  color: #99CC33;
+  padding-left: 1.5rem;
+  top: 10px;
+}
